rename make_delays into make_basic
