+++ TOC.md
@@ -88,8 +88,8 @@ Additional information, to be integrated into the list below:
 - [ ] make_arbitrary_grad
 - [x] make_arbitrary_rf                 *make_pulse.py*
 - [x] make_block_pulse                  *make_pulse.py*
-- [x] make_delay                        *make_delays.py*
-- [x] make_digital_output_pulse         *make_delays.py* **acts as delay in sim**
+- [x] make_delay                        *make_basic.py*
+- [x] make_digital_output_pulse         *make_basic.py* **acts as delay in sim**
 - [ ] make_extended_trapezoid
 - [ ] make_extended_trapezoid_area     (copied from pypulseq, not differentiable)
 - [x] make_gauss_pulse                  *make_pulse.py*
@@ -98,7 +98,7 @@ Additional information, to be integrated into the list below:
 - [-] make_slr                         **no sigpy support**
 - [-] make_sms                         **no sigpy support**
 - [x] make_trapezoid                    *make_grad.py*
-- [x] make_trigger                      *make_delays.py* **acts as delay in sim**
+- [x] make_trigger                      *make_basic.py* **acts as delay in sim**
 - [ ] points_to_waveform
 - [ ] rotate
 - [x] round_half_up                     *math.py*
